Fix compat with RubyInstaller-2.4 on Windows (#875)
Since RubyInstaller-2.4+ is bundled with MSYS2 and the libmariadbclient
can be installed per gemspec library dependency, it is easy to build
the mysql2 gem in Windows.

The MSYS2/MINGW dependency feature is documented here:
https://github.com/oneclick/rubyinstaller2/wiki/For-gem-developers#msys2-library-dependency

This also adds ruby-2.4 binaries, so that the mysql2 is still usabel as
a binary gem.

Fixes #861

The change in the spec is required for mariadbclient. It throws an error
if no query was executed.

Due to the stdcall convention on i686, the mysql_query() function check
fails, so that it is omitted, now.

Signed-off-by: Jun Aruga <jaruga@redhat.com>